pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm[toml]"]
build-backend = "setuptools.build_meta"

[project]
name = "git_ai_commit"
description = "AI-powered commit message generator that analyzes git diffs"
keywords = ["git", "commit", "ai", "openai", "diff", "commit message"]
license = { text = "MIT License" }
urls = { repository = "https://github.com/klokie/git-helper" }
authors = [{ name = "Daniel Grossfeld", email = "github@klokie.com" }]
requires-python = ">= 3.11"
dependencies = ["openai>=1.12.0", "python-dotenv~=1.0.0"]
dynamic = ["version", "readme"]

[tool.setuptools.dynamic]
readme = { file = ["README.md"] }

[project.scripts]
git_ai_commit = "git_ai_commit.cli:main"

[tool.setuptools]
packages = ["git_ai_commit"]
include-package-data = false

[tool.setuptools_scm]
